+++ com.unity.mobile.android-logcat/Editor/AndroidLogcatUtilities.cs
@@ -3,10 +3,11 @@ using System.IO;
 using System.Text.RegularExpressions;
 using UnityEngine;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Unity.Android.Logcat
 {
-    internal class AndroidLogcatUtilities
+    internal static class AndroidLogcatUtilities
     {
         internal static readonly string kAbiArm64 = "arm64-v8a";
         internal static readonly string kAbiArmV7 = "armeabi-v7a";
@@ -398,13 +399,12 @@ namespace Unity.Android.Logcat
         /// <param name="symbolPath"></param>
         /// <param name="libraryFile"></param>
         /// <returns></returns>
-        internal static string GetSymbolFile(string symbolPath, string libraryFile)
+        internal static string GetSymbolFile(string symbolPath, string libraryFile, string[] extensionsToTry)
         {
             var fullPath = Path.GetFullPath(Path.Combine(symbolPath, libraryFile));
             if (File.Exists(fullPath))
                 return fullPath;
 
-            var extensionsToTry = new[] { ".sym.so", ".dbg.so" };
             foreach (var e in extensionsToTry)
             {
                 // Try sym.so extension
@@ -416,7 +416,7 @@ namespace Unity.Android.Logcat
             return string.Empty;
         }
 
-        internal static string GetSymbolFile(IReadOnlyList<ReordableListItem> symbolPaths, string abi, string libraryFile)
+        internal static string GetSymbolFile(IReadOnlyList<ReordableListItem> symbolPaths, string abi, string libraryFile, string[] extensionsToTry)
         {
             foreach (var symbolPath in symbolPaths)
             {
@@ -425,12 +425,12 @@ namespace Unity.Android.Logcat
 
                 if (!string.IsNullOrEmpty(abi))
                 {
-                    var fileWithABI = GetSymbolFile(Path.Combine(symbolPath.Name, abi), libraryFile);
+                    var fileWithABI = GetSymbolFile(Path.Combine(symbolPath.Name, abi), libraryFile, extensionsToTry);
                     if (!string.IsNullOrEmpty(fileWithABI))
                         return fileWithABI;
                 }
 
-                var file = GetSymbolFile(symbolPath.Name, libraryFile);
+                var file = GetSymbolFile(symbolPath.Name, libraryFile, extensionsToTry);
                 if (!string.IsNullOrEmpty(file))
                     return file;
             }
@@ -526,6 +526,11 @@ namespace Unity.Android.Logcat
             {
                 return false;
             }
+        }
+
+        internal static string[] GetEnabledValues(this IReadOnlyList<ReordableListItem> list)
+        {
+            return list.Where(i => i.Enabled).Select(i => i.Name).ToArray();
         }
     }
 }